Scale mouse wheel scroll speed by zoom factor
Keeps the same number of text lines scrolled per wheel tick regardless
of zoom level, preventing overly fast scrolling when zoomed out.

+++ RaisinDocs/ScrollController.cs
@@ -46,7 +46,7 @@ internal class ScrollController
 
     internal void CancelSmooth() => _smoother.Cancel();
 
-    internal void HandleWheel(int delta)
+    internal void HandleWheel(double delta)
     {
         if (_smoother.IsAnimating)
         {
